requirements.txt
moviepy
Pillow
google-genai
python-dotenv